refactor(arch): break the four import cycles under .agents/scripts; ratchet allowlist 4 → 0 (refs #3995)
Break all four live import cycles flagged by the check:arch gate (#3991)
and shrink baselines/arch-cycles.json to empty so the dependency graph is
acyclic by construction. Module locations and import paths only — no
exported function's name, signature, or behaviour changes; all call sites
and tests updated in the same change, no shims.

- Cycles 1+2 (baselines/metrics cluster): extract transpileIfNeeded /
  resolveTsTranspilerVersion into the leaf lib/transpile.js (imported by
  engine, utils, workers, native, crap-utils); relocate the
  baseline-read/write arm (getBaseline / saveBaseline) out of
  maintainability-utils.js into lib/baselines/maintainability-baseline-io.js
  (reader leaf) and lib/baselines/maintainability-baseline-save.js (writer).
  Flow is now one-way: kernel → kinds → utils, writer → kernel inside the
  package.
- Cycle 3 (ticketing): extract the single-ticket mutators
  (transitionTicketState, toggleTasklistCheckbox, postStructuredComment)
  into the leaf ticketing/transition.js; bulk.js depends downward on it,
  and state.js injects bulk's upward cascade via registerCascadeRunner so
  transition.js never imports bulk.js. state.js re-exports the moved
  primitives so the facade and all consumers are unchanged.
- Cycle 4 (validation): relocate computeStoryReachability into the leaf
  lib/orchestration/story-reachability.js; both file-assumptions.js and
  ticket-validator-conflicts.js import it from below.
- Re-attribute the refresh-service invariant allowlist entry for
  saveBaseline from maintainability-utils.js to its new home.

Verify: npm run lint (check:arch green, empty allowlist), npm test
(8778 pass / 0 fail), node .agents/scripts/check-baselines.js (exit 0),
zero phantom baseline rows.
